[CPU] Added jit_snippets_call_args.hpp
