+++ src/plugins/intel_cpu/src/emitters/snippets/x64/jit_kernel_emitter.hpp
@@ -6,69 +6,12 @@
 
 #include "emitters/plugin/x64/jit_emitter.hpp"
 
+#include "emitters/snippets/jit_snippets_call_args.hpp"
 #include "jit_container_emitter.hpp"
 
 
 namespace ov {
 namespace intel_cpu {
-
-#define SNIPPETS_MAX_SNIPPETS_DIMS 12
-#define SNIPPETS_MAX_HARNESS_DIMS 5
-#define SNIPPETS_MAX_TILE_RANK 2
-#define SNIPPETS_DYNAMIC_MASTER_SHAPE_RANK 6
-
-#define GET_OFF(field) offsetof(jit_snippets_call_args, field)
-#define GET_OFF_LOOP_ARGS(field) offsetof(jit_snippets_call_args::loop_args_t, field)
-
-struct amx_tile_config_t {
-    size_t M = 0;
-    size_t K = 0;
-    size_t N = 0;
-};
-
-struct jit_snippets_call_args {
-    struct loop_args_t;
-
-    jit_snippets_call_args() = default;
-    ~jit_snippets_call_args();
-
-    void register_loops(const std::vector<loop_args_t>& loops);
-
-    const void *src_ptrs[SNIPPETS_MAX_SNIPPETS_DIMS] = {};
-    void *dst_ptrs[SNIPPETS_MAX_SNIPPETS_DIMS] = {};
-    void *buffer_scratchpad_ptr = nullptr;
-    // Note: Ideally loop_args must be private, since we manage this pointer manually.
-    // However, standard-layout class definition (to use offset_of) requires the same access specifier
-    // for all non-static data members. So we can keep them public or friend all control-flow emitters
-    int32_t num_loops = 0;
-    loop_args_t* loop_args = nullptr;
-    amx_tile_config_t amx_tile_config;
-};
-
-struct jit_snippets_call_args::loop_args_t {
-    friend class jit_loop_begin_dynamic_emitter;
-    friend class jit_loop_end_dynamic_emitter;
-
-    loop_args_t() = default;
-    loop_args_t(int64_t work_amount, const std::vector<int64_t>& ptr_increments, const std::vector<int64_t>& finalization_offsets);
-    loop_args_t(const loop_args_t& other);
-    ~loop_args_t();
-
-    loop_args_t& operator=(loop_args_t other);
-    friend void swap(loop_args_t& first, loop_args_t& second);
-
-private:
-    void init_pointers_and_copy_data(const int64_t num_elements, const int64_t* ptr_increments, const int64_t* finalization_offsets);
-
-    int64_t m_work_amount = 0;
-    int64_t m_num_data_ptrs = 0;
-    int64_t* m_ptr_increments = nullptr;
-    int64_t* m_finalization_offsets = nullptr;
-};
-
-struct jit_snippets_compile_args {
-    size_t parallel_executor_ndims = 1;
-};
 
 ///
 /// \brief    Kernel is the only entry point to Codogen Jit compilation. Kernel perform abstract-to-physical register
